First commited as before_33223_rev3

+++ lttng/org.eclipse.linuxtools.lttng2.kernel.core.tests/src/org/eclipse/linuxtools/lttng2/kernel/core/tests/stateprovider/PartialStateSystemTest.java
@@ -67,8 +67,12 @@ public class PartialStateSystemTest extends StateSystemTest {
      */
     @After
     public void cleanup() {
-        module.close();
-        stateFile.delete();
+        if (module != null) {
+            module.close();
+        }
+        if (stateFile != null) {
+            stateFile.delete();
+        }
     }
 
     /**
